requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
spacy==3.7.2
transformers==4.33.0
torch==2.0.1
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.9.0
nltk==3.8.1
plotly==5.15.0
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2